Move the oslc-ui-model package under server
