+++ pom.xml
@@ -76,7 +76,6 @@
   <modules>
     <module>core/oslc4j-core-build</module>
     <module>domains/oslc-domains</module>
-    <module>domains/ui-classes</module>
     <module>client</module>
     <module>server</module>
