+++ server/oauth-webapp/pom.xml
@@ -10,7 +10,7 @@
     <groupId>org.eclipse.lyo.server</groupId>
     <artifactId>oauth-webapp</artifactId>
     <packaging>war</packaging>
-    <name>Lyo :: OAuth :: Webapp</name>
+    <name>Lyo :: Server :: OAuth Webapp</name>
     <description>A template webapp for implementing an OAuth provider.</description>
 
     <properties>
